requirements.txt
playwright
beautifulsoup4
google-generativeai
Pillow
flask
twilio
requests
